+++ pkg/grpcfactory/client.go
@@ -12,7 +12,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package provisioner
+package grpcfactory
 
 import (
 	"context"
