+++ pkg/grpcfactory/server.go
@@ -3,30 +3,36 @@ Copyright 2024 Scality, Inc.
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
+
 	http://www.apache.org/licenses/LICENSE-2.0
+
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package provisioner
+package grpcfactory
+
 import (
 	"context"
 	"fmt"
 	"net"
 	"net/url"
+
 	"github.com/pkg/errors"
 	"google.golang.org/grpc"
 	"k8s.io/klog/v2"
 	cosi "sigs.k8s.io/container-object-storage-interface-spec"
 )
+
 type COSIProvisionerServer struct {
 	address           string
 	identityServer    cosi.IdentityServer
 	provisionerServer cosi.ProvisionerServer
-	listenOpts []grpc.ServerOption
+	listenOpts        []grpc.ServerOption
 }
+
 func (s *COSIProvisionerServer) Run(ctx context.Context) error {
 	addr, err := url.Parse(s.address)
 	if err != nil {
